# artifacts/openapi/wecms-api-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeCMS API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/system/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "GET /api/v1/system/users",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PagedUserSummary"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "POST /api/v1/system/users",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:create",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "create"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "DELETE /api/v1/system/users/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:delete",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "delete"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "GET /api/v1/system/users/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDetailDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:detail"
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "PUT /api/v1/system/users/{id:long}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserMutationResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:update",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "update"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/disable": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "POST /api/v1/system/users/{id:long}/disable",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:disable",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "disable"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/enable": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "POST /api/v1/system/users/{id:long}/enable",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:enable",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "enable"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/positions": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "PUT /api/v1/system/users/{id:long}/positions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserPositionsRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:assign-position",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "assign-position"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/reset-2fa": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "POST /api/v1/system/users/{id:long}/reset-2fa",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserTwoFactorRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:reset-2fa",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "reset-2fa"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/reset-password": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "POST /api/v1/system/users/{id:long}/reset-password",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:reset-password",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "reset-password"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
    "/api/v1/system/users/{id:long}/roles": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "PUT /api/v1/system/users/{id:long}/roles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResult"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserRolesRequest"
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-wecms-module": "identity",
        "x-wecms-permission": "sys:user:assign-role",
        "x-wecms-audit": {
          "module": "identity",
          "resource": "users",
          "action": "assign-role"
        },
        "x-wecms-rate-limit": "admin_write_policy"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResult": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "nullable": true
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          },
          "fieldErrors": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "Object": {
        "type": "object",
        "additionalProperties": true
      },
...
      "PagedUserSummary": {
        "type": "object",
        "required": [
          "records",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummaryDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UserSummaryDto": {
        "type": "object",
        "required": [
          "id",
          "username",
          "displayName",
          "status",
          "isSuperAdmin",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "deptId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "isSuperAdmin": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserDetailDto": {
        "type": "object",
        "required": [
          "id",
          "username",
          "displayName",
          "status",
          "isSuperAdmin",
          "permissionVersion",
          "roleIds",
          "positionIds",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "deptId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "isSuperAdmin": {
            "type": "boolean"
          },
          "permissionVersion": {
            "type": "integer",
            "format": "int64"
          },
          "lastLoginAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "positionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username",
          "displayName",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "deptId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "roleIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "positionIds": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "required": [
          "displayName"
        ],
        "properties": {
          "displayName": {
            "type": "string"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "deptId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "ResetUserPasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
      "ResetUserTwoFactorRequest": {
        "type": "object",
        "required": [
          "reason"
        ],
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "AssignUserRolesRequest": {
        "type": "object",
        "required": [
          "roleIds"
        ],
        "properties": {
          "roleIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "AssignUserPositionsRequest": {
        "type": "object",
        "required": [
          "positionIds"
        ],
        "properties": {
          "positionIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "UserMutationResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}